Simplify es6 setup
