+++ package.json
@@ -16,7 +16,6 @@
   },
   "license": "MIT",
   "main": "lib/scripts/index.js",
-  "module": "lib/scripts/index.esm.js",
   "bin": {
     "styleguidist": "lib/bin/styleguidist.js"
   },
